Bring the minute archive up to date
There was no contact with the Tech Priest before me, so these are mostly html files, with the headers corrected, dates set, and so on. As there will be no Tech Priest next year, a Google Drive link has been put in, so the secretary can upload minutes directly (as they seem to have been doing for a few years now).